MSFT_Agent_Framework/requirements.txt
agent-framework
azure-ai-projects
azure-identity
azure-monitor-opentelemetry